third_party/WebKit/LayoutTests/http/tests/inspector/bindings/navigator-main-frame-navigated.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="./bindings-test.js"></script>

<script>

async function test()
{
    var sourcesNavigator = new Sources.SourcesNavigatorView();
    sourcesNavigator.show(UI.inspectorView.element);

    InspectorTest.markStep('dumpInitialNavigator');
    InspectorTest.dumpNavigatorView(sourcesNavigator, false);

    InspectorTest.markStep('attachFrame');
    await Promise.all([
        InspectorTest.attachFrame('frame', './resources/sourcemap-frame.html'),
        InspectorTest.waitForSourceMap('sourcemap-script.js.map'),
        InspectorTest.waitForSourceMap('sourcemap-style.css.map'),
    ]);
    InspectorTest.dumpNavigatorView(sourcesNavigator, false);

    InspectorTest.markStep('navigateMainFrame');
    var url = "http://127.0.0.1:8000/inspector/bindings/resources/empty-page.html";
    await InspectorTest.navigatePromise(url);
    InspectorTest.dumpNavigatorView(sourcesNavigator, false);

    InspectorTest.markStep('navigateBack');
    var originalUrl = "http://127.0.0.1:8000/inspector/bindings/navigator-main-frame-navigated.html";
    await InspectorTest.navigatePromise(originalUrl);
    InspectorTest.completeTest();
}

</script>

</head>
<body onload="runTest()">
<p>
Verify that navigator properly handles main frame navigated event.
</p>
</body>
</html>
