third_party/WebKit/LayoutTests/http/tests/inspector/bindings/navigator-main-frame-navigated-expected.txt

Verify that navigator properly handles main frame navigated event.


Running: dumpInitialNavigator
top
  127.0.0.1:8000
    inspector
      bindings
        navigator-main-frame-navigated.html
        bindings-test.js
      inspector-test.js

Running: attachFrame
top
  127.0.0.1:8000
    inspector
      bindings
        navigator-main-frame-navigated.html
        bindings-test.js
      inspector-test.js
  frame (magic-frame.html)
    127.0.0.1:8000
      inspector/bindings/resources
        magic-frame.html
        magic-script.js
        magic-style.css

Running: navigateMainFrame
Page reloaded.
top
  localhost:8000
    inspector/bindings/resources
      empty-page.html

Running: navigateBack
Page reloaded.
